Move Key stage performance show/hide logic to data model
